refactor(#3335): rename modules
