+++ deps/wazuh_testing/wazuh_testing/modules/aws/event_monitor.py
